+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Anthracite now also superheats blaze burners @Pyritie
 - Changed LV chem reactor quest to work with any container of rubber instead of just a rubber bucket (#1116) @ofoxsmith
 - Added perfect overclock to the electric greenhouse so it scales better with the pyro oven @Pyritie
+- Increased durability of the blue steel diving suit to match blue steel armor @Pyritie
 ### Bug fixes
 - Fixed issue where "melts into" tooltips and certain timers on Jade were missing (#1102) @Xikaro
 - Improved block ordering logic with AoE mining tools to reduce incorrect cave-ins @voidstar240
@@ -27,7 +28,7 @@
 - Fixed kaolin clay duping (#1099) @SpicyNoodle5
 - Fixed TFC sugar recipes so they can now use any sugar (#1099) @SpicyNoodle5
 - Add the create toolbox to the list of unsortable inventories, fixing a duping exploit (#1120) @esotericist
-- Fixed assembler recipe for item silos
+- Fixed assembler recipe for item silos @Pyritie
 
 ## [0.9.9] - 30.05.2025
 ### Changes